+++ src/cmd/cmd.rs
@@ -135,6 +135,7 @@ pub enum InitHook {
 #[derive(ValueEnum, Clone, Debug)]
 pub enum InitShell {
     Bash,
+    Cmd,
     Elvish,
     Fish,
     Nushell,