+++ src/util.rs
@@ -291,44 +291,77 @@ pub fn resolve_path(path: impl AsRef<Path>) -> Result<PathBuf> {
             }
         }
 
-        fn get_drive_path(drive_letter: u8) -> PathBuf {
-            format!(r"{}:\", drive_letter as char).into()
+        fn get_drive_prefix_path(drive_letter: u8) -> PathBuf {
+            format!(r"{}:\", patch_drive_letter(drive_letter)).into()
         }
 
-        fn get_drive_relative(drive_letter: u8) -> Result<PathBuf> {
+        fn get_drive_relative_path(drive_letter: u8) -> Result<PathBuf> {
             let path = current_dir()?;
             if Some(drive_letter) == get_drive_letter(&path) {
-                return Ok(path);
+                return Ok(patch_drive_prefix(path));
             }
 
-            if let Some(path) = env::var_os(format!("={}:", drive_letter as char)) {
-                return Ok(path.into());
+            if let Some(path) = env::var_os(format!("={}:", patch_drive_letter(drive_letter))) {
+                return Ok(patch_drive_prefix(path.into()));
             }
 
-            let path = get_drive_path(drive_letter);
+            let path = get_drive_prefix_path(drive_letter);
             Ok(path)
+        }
+
+        #[inline(always)]
+        fn patch_drive_letter(drive_letter: u8) -> char {
+            drive_letter.to_ascii_uppercase() as char
+        }
+
+        // https://github.com/rust-lang/rust-analyzer/pull/14689
+        fn patch_drive_prefix(path: PathBuf) -> PathBuf {
+            let mut components = path.components();
+
+            match components.next() {
+                Some(Component::Prefix(prefix)) => {
+                    let prefix = match prefix.kind() {
+                        Prefix::Disk(drive_letter) => {
+                            format!(r"{}:", patch_drive_letter(drive_letter))
+                        }
+                        Prefix::VerbatimDisk(drive_letter) => {
+                            format!(r"\\?\{}:", patch_drive_letter(drive_letter))
+                        }
+                        _ => return path,
+                    };
+
+                    let mut path = PathBuf::default();
+                    path.push(prefix);
+                    path.extend(components);
+                    path
+                }
+                _ => path,
+            }
         }
 
         match components.peek() {
             Some(Component::Prefix(prefix)) => match prefix.kind() {
                 Prefix::Disk(drive_letter) => {
-                    let disk = components.next().unwrap();
+                    components.next();
                     if components.peek() == Some(&Component::RootDir) {
-                        let root = components.next().unwrap();
-                        stack.push(disk);
-                        stack.push(root);
+                        components.next();
+                        base_path = get_drive_prefix_path(drive_letter);
                     } else {
-                        base_path = get_drive_relative(drive_letter)?;
-                        stack.extend(base_path.components());
+                        base_path = get_drive_relative_path(drive_letter)?;
                     }
+
+                    stack.extend(base_path.components());
                 }
                 Prefix::VerbatimDisk(drive_letter) => {
                     components.next();
                     if components.peek() == Some(&Component::RootDir) {
                         components.next();
+                        base_path = get_drive_prefix_path(drive_letter);
+                    } else {
+                        // Verbatim prefix without a root component? Likely not a legal path
+                        bail!("illegal path: {}", path.display());
                     }
 
-                    base_path = get_drive_path(drive_letter);
                     stack.extend(base_path.components());
                 }
                 _ => bail!("invalid path: {}", path.display()),
@@ -340,7 +373,7 @@ pub fn resolve_path(path: impl AsRef<Path>) -> Result<PathBuf> {
                 let drive_letter = get_drive_letter(&current_dir).with_context(|| {
                     format!("could not get drive letter: {}", current_dir.display())
                 })?;
-                base_path = get_drive_path(drive_letter);
+                base_path = get_drive_prefix_path(drive_letter);
                 stack.extend(base_path.components());
             }
             _ => {
@@ -376,4 +409,39 @@ pub fn resolve_path(path: impl AsRef<Path>) -> Result<PathBuf> {
 pub fn to_lowercase(s: impl AsRef<str>) -> String {
     let s = s.as_ref();
     if s.is_ascii() { s.to_ascii_lowercase() } else { s.to_lowercase() }
+}
+
+#[cfg(test)]
+#[cfg(windows)]
+mod tests_win {
+    use std::path::PathBuf;
+
+    use rstest::rstest;
+
+    #[rstest]
+    #[case(r"c:\", r"C:\")]
+    #[case(r"C:\", r"C:\")]
+    #[case(r"c:\\.", r"C:\")]
+    #[case(r"c:\..", r"C:\")]
+    #[case(r"C:\.\.", r"C:\")]
+    #[case(r"\\?\C:\", r"C:\")]
+    #[case(r"\\?\c:\", r"C:\")]
+    #[case(r"\\?\C:\\\", r"C:\")]
+    #[case(r"\\?\c:\\.\", r"C:\")]
+    #[case(r"c:\Windows", r"C:\Windows")]
+    #[case(r"C:\WINDOWS", r"C:\WINDOWS")]
+    #[case(r"c:\\\Windows\.", r"C:\Windows")]
+    #[case(r"C:\$WinREAgent", r"C:\$WinREAgent")]
+    #[case(r"\\?\c:\\Windows\\.", r"C:\Windows")]
+    #[case(r"\\?\c:\..\.\windows", r"C:\windows")]
+    #[case(r"c:\Windows\System32\.", r"C:\Windows\System32")]
+    #[case(r"c:\WINDOWS\..\..\Windows", r"C:\Windows")]
+    #[case(r"c:\Windows\..\.\.\..\Temp\..\tmp", r"C:\tmp")]
+    #[case(r"c:\.\Windows\..\..\Program Files", r"C:\Program Files")]
+    #[case(r"\\?\C:\\$WinREAgent\\..\Program Files\.", r"C:\Program Files")]
+    fn resolve_path(#[case] absolute_path: &str, #[case] normalized_form: &str) {
+        let path = PathBuf::from(absolute_path);
+        let resolved_path = super::resolve_path(path).unwrap();
+        assert_eq!(resolved_path.to_str().unwrap(), normalized_form);
+    }
 }