+++ src/cmd/query.rs
@@ -1,10 +1,10 @@
 use std::io::{self, Write};
 
-use anyhow::{bail, Context, Result};
+use anyhow::{Context, Result};
 
 use crate::cmd::{Query, Run};
 use crate::config;
-use crate::db::{Database, Epoch, Stream};
+use crate::db::{Database, Epoch, Stream, StreamOptions};
 use crate::error::BrokenPipeHandler;
 use crate::util::{self, Fzf, FzfChild};
 
@@ -18,59 +18,75 @@ impl Run for Query {
 impl Query {
     fn query(&self, db: &mut Database) -> Result<()> {
         let now = util::current_time()?;
-        let mut stream = self.get_stream(db, now);
+        let mut stream = self.get_stream(db, now)?;
 
         if self.interactive {
-            let mut fzf = Self::get_fzf()?;
-            let selection = loop {
-                match stream.next() {
-                    Some(dir) => {
-                        if let Some(selection) = fzf.write(dir, now)? {
-                            break selection;
-                        }
-                    }
-                    None => break fzf.wait()?,
-                }
-            };
-
-            if self.score {
-                print!("{selection}");
-            } else {
-                let path = selection.get(7..).context("could not read selection from fzf")?;
-                print!("{path}");
-            }
+            self.query_interactive(&mut stream, now)
         } else if self.list {
-            let handle = &mut io::stdout().lock();
-            while let Some(dir) = stream.next() {
-                let dir = if self.score { dir.display().with_score(now) } else { dir.display() };
-                writeln!(handle, "{dir}").pipe_exit("stdout")?;
-            }
+            self.query_list(&mut stream, now)
         } else {
-            let handle = &mut io::stdout();
-            let Some(dir) = stream.next() else {
-                bail!(if stream.did_exclude() {
-                    "you are already in the only match"
-                } else {
-                    "no match found"
-                });
-            };
-            let dir = if self.score { dir.display().with_score(now) } else { dir.display() };
-            writeln!(handle, "{dir}").pipe_exit("stdout")?;
+            self.query_first(&mut stream, now)
         }
+    }
 
+    fn query_interactive(&self, stream: &mut Stream, now: Epoch) -> Result<()> {
+        let mut fzf = Self::get_fzf()?;
+        let selection = loop {
+            match stream.next() {
+                Some(dir) if Some(dir.path.as_ref()) == self.exclude.as_deref() => continue,
+                Some(dir) => {
+                    if let Some(selection) = fzf.write(dir, now)? {
+                        break selection;
+                    }
+                }
+                None => break fzf.wait()?,
+            }
+        };
+
+        if self.score {
+            print!("{selection}");
+        } else {
+            let path = selection.get(7..).context("could not read selection from fzf")?;
+            print!("{path}");
+        }
         Ok(())
     }
 
-    fn get_stream<'a>(&self, db: &'a mut Database, now: Epoch) -> Stream<'a> {
-        let mut stream = db.stream(now).with_keywords(&self.keywords);
+    fn query_list(&self, stream: &mut Stream, now: Epoch) -> Result<()> {
+        let handle = &mut io::stdout().lock();
+        while let Some(dir) = stream.next() {
+            if Some(dir.path.as_ref()) == self.exclude.as_deref() {
+                continue;
+            }
+            let dir = if self.score { dir.display().with_score(now) } else { dir.display() };
+            writeln!(handle, "{dir}").pipe_exit("stdout")?;
+        }
+        Ok(())
+    }
+
+    fn query_first(&self, stream: &mut Stream, now: Epoch) -> Result<()> {
+        let handle = &mut io::stdout();
+
+        let mut dir = stream.next().context("no match found")?;
+        while Some(dir.path.as_ref()) == self.exclude.as_deref() {
+            dir = stream.next().context("you are already in the only match")?;
+        }
+
+        let dir = if self.score { dir.display().with_score(now) } else { dir.display() };
+        writeln!(handle, "{dir}").pipe_exit("stdout")
+    }
+
+    fn get_stream<'a>(&self, db: &'a mut Database, now: Epoch) -> Result<Stream<'a>> {
+        let mut options = StreamOptions::new(now)
+            .with_keywords(self.keywords.iter().map(|s| s.as_str()))
+            .with_exclude(config::exclude_dirs()?);
         if !self.all {
             let resolve_symlinks = config::resolve_symlinks();
-            stream = stream.with_exists(resolve_symlinks);
+            options = options.with_exists(true).with_resolve_symlinks(resolve_symlinks);
         }
-        if let Some(path) = &self.exclude {
-            stream = stream.with_exclude(path);
-        }
-        stream
+
+        let stream = Stream::new(db, options);
+        Ok(stream)
     }
 
     fn get_fzf() -> Result<FzfChild> {