+++ src/db/stream.rs
@@ -2,77 +2,41 @@ use std::iter::Rev;
 use std::ops::Range;
 use std::{fs, path};
 
+use glob::Pattern;
+
 use crate::db::{Database, Dir, Epoch};
 use crate::util::{self, MONTH};
 
 pub struct Stream<'a> {
-    // State
     db: &'a mut Database,
     idxs: Rev<Range<usize>>,
-    did_exclude: bool,
-
-    // Configuration
-    keywords: Vec<String>,
-    check_exists: bool,
-    expire_below: Epoch,
-    resolve_symlinks: bool,
-    exclude_path: Option<String>,
+    options: StreamOptions,
 }
 
 impl<'a> Stream<'a> {
-    pub fn new(db: &'a mut Database, now: Epoch) -> Self {
-        db.sort_by_score(now);
+    pub fn new(db: &'a mut Database, options: StreamOptions) -> Self {
+        db.sort_by_score(options.now);
         let idxs = (0..db.dirs().len()).rev();
-
-        // If a directory is deleted and hasn't been used for 3 months, delete
-        // it from the database.
-        let expire_below = now.saturating_sub(3 * MONTH);
-
-        Stream {
-            db,
-            idxs,
-            did_exclude: false,
-            keywords: Vec::new(),
-            check_exists: false,
-            expire_below,
-            resolve_symlinks: false,
-            exclude_path: None,
-        }
-    }
-
-    pub fn with_exclude(mut self, path: impl Into<String>) -> Self {
-        self.exclude_path = Some(path.into());
-        self
-    }
-
-    pub fn with_exists(mut self, resolve_symlinks: bool) -> Self {
-        self.check_exists = true;
-        self.resolve_symlinks = resolve_symlinks;
-        self
-    }
-
-    pub fn with_keywords(mut self, keywords: &[impl AsRef<str>]) -> Self {
-        self.keywords = keywords.iter().map(util::to_lowercase).collect();
-        self
+        Stream { db, idxs, options }
     }
 
     pub fn next(&mut self) -> Option<&Dir> {
         while let Some(idx) = self.idxs.next() {
             let dir = &self.db.dirs()[idx];
 
-            if !self.matches_keywords(&dir.path) {
+            if !self.filter_by_keywords(&dir.path) {
                 continue;
             }
 
-            if !self.matches_exists(&dir.path) {
-                if dir.last_accessed < self.expire_below {
+            if !self.filter_by_exclude(&dir.path) {
+                self.db.swap_remove(idx);
+                continue;
+            }
+
+            if !self.filter_by_exists(&dir.path) {
+                if dir.last_accessed < self.options.ttl {
                     self.db.swap_remove(idx);
                 }
-                continue;
-            }
-
-            if Some(dir.path.as_ref()) == self.exclude_path.as_deref() {
-                self.did_exclude = true;
                 continue;
             }
 
@@ -83,20 +47,8 @@ impl<'a> Stream<'a> {
         None
     }
 
-    pub fn did_exclude(&self) -> bool {
-        self.did_exclude
-    }
-
-    fn matches_exists(&self, path: &str) -> bool {
-        if !self.check_exists {
-            return true;
-        }
-        let resolver = if self.resolve_symlinks { fs::symlink_metadata } else { fs::metadata };
-        resolver(path).map(|m| m.is_dir()).unwrap_or_default()
-    }
-
-    fn matches_keywords(&self, path: &str) -> bool {
-        let (keywords_last, keywords) = match self.keywords.split_last() {
+    fn filter_by_keywords(&self, path: &str) -> bool {
+        let (keywords_last, keywords) = match self.options.keywords.split_last() {
             Some(split) => split,
             None => return true,
         };
@@ -121,6 +73,77 @@ impl<'a> Stream<'a> {
         }
 
         true
+    }
+
+    fn filter_by_exclude(&self, path: &str) -> bool {
+        !self.options.exclude.iter().any(|pattern| pattern.matches(path))
+    }
+
+    fn filter_by_exists(&self, path: &str) -> bool {
+        if !self.options.exists {
+            return true;
+        }
+        let resolver =
+            if self.options.resolve_symlinks { fs::metadata } else { fs::symlink_metadata };
+        resolver(path).map(|metadata| metadata.is_dir()).unwrap_or_default()
+    }
+}
+
+pub struct StreamOptions {
+    /// The current time.
+    now: Epoch,
+
+    /// Only directories matching these keywords will be returned.
+    keywords: Vec<String>,
+
+    /// Directories that match any of these globs will be lazily removed.
+    exclude: Vec<Pattern>,
+
+    /// Directories will only be returned if they exist on the filesystem.
+    exists: bool,
+
+    /// Whether to resolve symlinks when checking if a directory exists.
+    resolve_symlinks: bool,
+
+    /// Directories that do not exist and haven't been accessed since TTL will
+    /// be lazily removed.
+    ttl: Epoch,
+}
+
+impl StreamOptions {
+    pub fn new(now: Epoch) -> Self {
+        StreamOptions {
+            now,
+            keywords: Vec::new(),
+            exclude: Vec::new(),
+            exists: false,
+            resolve_symlinks: false,
+            ttl: now.saturating_sub(3 * MONTH),
+        }
+    }
+
+    pub fn with_keywords<I>(mut self, keywords: I) -> Self
+    where
+        I: IntoIterator,
+        I::Item: AsRef<str>,
+    {
+        self.keywords = keywords.into_iter().map(util::to_lowercase).collect();
+        self
+    }
+
+    pub fn with_exclude(mut self, exclude: Vec<Pattern>) -> Self {
+        self.exclude = exclude;
+        self
+    }
+
+    pub fn with_exists(mut self, exists: bool) -> Self {
+        self.exists = exists;
+        self
+    }
+
+    pub fn with_resolve_symlinks(mut self, resolve_symlinks: bool) -> Self {
+        self.resolve_symlinks = resolve_symlinks;
+        self
     }
 }
 
@@ -154,7 +177,8 @@ mod tests {
     #[case(&["/foo/", "/bar"], "/foo/baz/bar", true)]
     fn query(#[case] keywords: &[&str], #[case] path: &str, #[case] is_match: bool) {
         let db = &mut Database::new(PathBuf::new(), Vec::new(), |_| Vec::new(), false);
-        let stream = Stream::new(db, 0).with_keywords(keywords);
-        assert_eq!(is_match, stream.matches_keywords(path));
+        let options = StreamOptions::new(0).with_keywords(keywords.iter());
+        let stream = Stream::new(db, options);
+        assert_eq!(is_match, stream.filter_by_keywords(path));
     }
 }