+++ CHANGELOG.md
@@ -21,6 +21,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.9.2] - 2023-08-04
 
+- Support for `cmd.exe` (experimental).
+
 ### Added
 
 - Short option `-a` for `zoxide query --all`.