+++ CHANGELOG.md
@@ -12,7 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - fish: detect infinite loop when using `alias cd=z`.
-- fish / PowerShell: handle queries that look like args (e.g. `z -x`).
+- fish / Nushell / PowerShell: handle queries that look like args (e.g. `z -x`).
 
 ## [0.9.4] - 2024-02-21